requirements.txt
beautifulsoup4==4.12.3
fastapi==0.115.5
openai==1.54.3
httpx==0.27.2
python-dotenv==1.0.1
pydantic==2.9.2
requests==2.32.3
uvicorn[standard]==0.32.1
langchain==0.3.0
langchain-openai==0.2.0
langchain-community==0.3.0
tavily-python==0.5.0
duckduckgo-search==6.1.0

